backend/requirements.txt
Flask
Flask-RESTful
Flask-SQLAlchemy
mysqlclient
bcrypt
PyJWT
python-dateutil
json-delta
nosetests
